REFACTOR/#21: app 모듈 수정
