--- app/src/main/java/com/sopt/now/compose/container/repository/UserRepository.kt
@@ -1,11 +1,0 @@
-package com.sopt.now.compose.container.repository
-
-import com.sopt.now.compose.models.User
-
-interface UserRepository {
-    fun getUserProfile(): User?
-    fun setUserProfile(user: User)
-    fun getUserId(): String
-    fun setUserId(userId: String)
-}
-
