+++ app/src/main/java/com/sopt/now/compose/di/RepositoryModule.kt
@@ -1,13 +1,12 @@
 package com.sopt.now.compose.di
 
-import com.sopt.now.compose.container.impl.AuthRepositoryImpl
-import com.sopt.now.compose.container.impl.FollowerRepositoryImpl
-import com.sopt.now.compose.container.impl.MemberRepositoryImpl
-import com.sopt.now.compose.container.impl.UserRepositoryImpl
-import com.sopt.now.compose.container.repository.AuthRepository
-import com.sopt.now.compose.container.repository.FollowerRepository
-import com.sopt.now.compose.container.repository.MemberRepository
-import com.sopt.now.compose.container.repository.UserRepository
+
+import com.sopt.now.compose.data.repositoryimpl.AuthRepositoryImpl
+import com.sopt.now.compose.data.repositoryimpl.FollowerRepositoryImpl
+import com.sopt.now.compose.data.repositoryimpl.HomeRepositoryImpl
+import com.sopt.now.compose.domain.repository.AuthRepository
+import com.sopt.now.compose.domain.repository.FollowerRepository
+import com.sopt.now.compose.domain.repository.HomeRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -28,9 +27,5 @@ abstract class RepositoryModule {
 
     @Binds
     @Singleton
-    abstract fun bindMemberRepository(memberRepositoryImpl: MemberRepositoryImpl): MemberRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindUserRepository(userRepositoryImpl: UserRepositoryImpl): UserRepository
+    abstract fun bindHomeRepository(homeRepositoryImpl: HomeRepositoryImpl): HomeRepository
 }
