--- app/src/main/java/com/sopt/now/compose/models/Follower.kt
@@ -1,18 +1,0 @@
-package com.sopt.now.compose.models
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Follower(
-    @SerialName("id")
-    val id:Int,
-    @SerialName("email")
-    val email:String,
-    @SerialName("first_name")
-    val firstName:String,
-    @SerialName("last_name")
-    val lastName:String,
-    @SerialName("avatar")
-    val avatar:String
-)
