--- app/src/main/java/com/sopt/now/compose/models/User.kt
@@ -1,10 +1,0 @@
-package com.sopt.now.compose.models
-
-import java.io.Serializable
-
-data class User(
-    var id: String = "",
-    var pw: String = "",
-    var nickName: String = "",
-    var phone: String = "",
-): Serializable
