--- app/src/main/java/com/sopt/now/compose/network/dto/RequestChangePasswordDto.kt
@@ -1,14 +1,0 @@
-package com.sopt.now.compose.network.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestChangePasswordDto(
-    @SerialName("previousPassword")
-    val previousPassword: String,
-    @SerialName("newPassword")
-    val newPassword: String,
-    @SerialName("newPasswordVerification")
-    val newPasswordVerification: String
-)
